Implemented constraints for decoding.
-- constraints are given in a file containing bitstrings on each
line. Each bitstring corresponds to an input sentence. A 1 means
that we can reorder a chunk. A 0 means we can't. The number of bits
corresponds to the number of chunks defined by chopping.
-- implemented chopping from file. The file contains the chopping
information. Each line corresponding to an input sentence contains
space separated integers. Each integer corresponds to the zero-based index
of the word starting the next chunk. The last integer is the sentence length.
-- updated chopping so that the vector of integers returned by the chop
method is never empty and always contains the sentence length as last element.