Bug fix.
We now assume that the input bag of words has start-of-sentence (SOS)
and end-of-sentence (EOS) markers. The code was not assuming that
because in the past some inputs were chopped manually and
some input bags corresponding to the first chop had an SOS marker
and some other bags did not have an SOS marker. Because of this,
when the input actually had the start of sentence marker,
the code allowed to extend the initial state with ngrams
that did not start with SOS. This could cause the output
lattice to be empty in case of pruning.